Refactor into versioned driverinterface
